+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.7.0 - Unreleased
+
+### Changed
+
+- adapt to the new Hammer API
+- use [`:mnesia.dirty_update_counter/3`](https://www.erlang.org/doc/apps/mnesia/mnesia.html#dirty_update_counter/3)
+- automatically create an in-memory table (no schema needed)
+- listen for cluster changes and replicate the in-memory table
+- require Elixir 1.15+ and OTP 24+
 
 ## 0.6.1 - 2024-03-29